Set up static ClientID at clientid.jsonld (#69)
This is useful mainly because the user will (hopefully) not have to click `Authorize` during sign-in every time.

Creates new environment variables:
`BASE_URL` - the app url without trailing slash that is used in ClientID
`REACT_APP_ENABLE_DEV_CLIENT_ID` - enable static ClientID in development

It's causing issues with Vercel previews. If you want to try and want it
to work, you need to use "branch url",
https://sleepy-bike-git-static-clientid-ohn.vercel.app/ in this case. If
it's going to cause issues, we can disable this feature in Vercel
altogether.