Remove solidweb.me from recommended pod providers (#93)
Solidweb.me has been deprecated and is probably going to reach end of
life at the end of 2024, see #91

We remove solidweb.me from the _Join_ list, but keep it in _Sign in_
list. We want people to be able to keep using it, but not register new
accounts with it.

Also fix a warning regarding missing keys in list.