Clarify the attribution and license of the project (#119)
This codebase was originally developed as part of the Open Hospitality Network (OHN) and has been subsequently relicensed to SolidCouch under the current license (MIT). Any remaining files that are still licensed under OHN retain their original license (MIT), and these files are clearly marked with the original OHN license.

This change is necessary to address missing consent with the relicensing.

Approaches:
- Include previous license in relevant files
- Remove code